Web XML file updated to take into account changes made during refactoring
Issue identified with Steam API : According to this thread linked in
issue #183 Valve is now rate limiting all XML data globally. That means
they're only accepting a given number of request in a specific time
interval. If that limit is reached all other requests will be blocked
(with HTTP status 503).

https://github.com/koraktor/steam-condenser/issues/194