Register updated to connect to mobile phone app
Username and password now being sent to the server, new users are
registered, existing users are checked and the appropriate boolean
values are return in JSON.

Passwords are now hashed using SHA encryption to ensure security.
However passwords are send in plain text over the internet to the
server. Additional work required to ensure passwords are sent to the
server securely.